Replace RequestLoadingIndicator with enhanced Loading component
Removed the RequestLoadingIndicator component and updated App.tsx to use the Loading component instead. The Loading component now accepts a loadingText prop for customizable messages and integrates API and dropdown loading states for improved user feedback.
